+++ src/main/java/com/appdynamics/extensions/haproxy/config/ProxyStats.java
@@ -11,8 +11,6 @@ import javax.xml.bind.annotation.XmlRootElement;
 public class ProxyStats {
     @XmlElement(name = "stat")
     private Stat stat;
-    @XmlElement(name = "proxy-servers")
-    private ProxyServerConfig proxyServerConfig;
 
     public Stat getStat() {
         return stat;
@@ -20,14 +18,6 @@ public class ProxyStats {
 
     public void setStat() {
         this.stat = stat;
-    }
-
-    public ProxyServerConfig getProxyServerConfig() {
-        return proxyServerConfig;
-    }
-
-    public void setProxyServerConfig(ProxyServerConfig proxyServerConfig) {
-        this.proxyServerConfig = proxyServerConfig;
     }
 
 }