+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+### 2.1.0 - Sep 04, 2018
+1. Moved the proxy-server config to config.yml
+2. Added regex support for pxname-svname
+3. No support for [.*] in regex
 ### 2.0.0 - June 21, 2018
 1. Moved HAProxy Monitor to 2.0.0 framework
 2. Moved all the metrics map/dictionary to metrics.xml